+++ build-sdist.sh
@@ -10,8 +10,10 @@ CHANGELOG="HISTORY.rst"
 echo "Creating HISTORY.rst"
 echo "Latest Changes" > $CHANGELOG
 echo "==============" >> $CHANGELOG
-echo "Compiled on: `date`" >> $CHANGELOG
-echo "\nThis file is auto-generated upon every new release. To review the latest commits in the master branch, please refer to: https://github.com/django-wiki/django-wiki/commits/master"
+echo "" >> $CHANGELOG
+echo "This file is auto-generated upon every new release. To review the latest commits in the master branch, please refer to: https://github.com/django-wiki/django-wiki/commits/master" >> $CHANGELOG
+echo "" >> $CHANGELOG
+echo "Compiled on: `date`::" >> $CHANGELOG
 echo "" >> $CHANGELOG
 git log --graph --pretty=format:'%h -%d %s (%cr) <%an>' --abbrev-commit | sed "s/^/    /" >> $CHANGELOG